[CallFrameInterface] Add methods `ConsumeArg()` and `CanConsumeArg()`.
At present, the `CallFrameInterface` (and, by extension, all TensorFlow functions that pass arguments via `ArgOp`) retains ownership of one reference on each argument tensor for the lifetime of the frame. This prevents buffer forwarding optimizations from being performed on arguments, which can lead to performance issues. However, in some cases (e.g. the loop variables in a `WhileOp`, the arguments to the map function in `Dataset.map()`) we know enough about the caller to be able to "move" the arguments into the function call.

This change adds the following:
* Methods called `CanConsumeArg(int index)` and `ConsumeArg(int index, Tensor* val)` to `CallFrameInterface` that allow the runtime to query the call frame for consumable arguments. The default implementations do not permit any arguments to be consumed, for backwards compatibility.
* An implementation of these methods in tf.data's "captured_function.cc", allowing arguments to `Dataset.map()` functions (and similar) to be consumed by the function call.
* A specialized `CallFrameInterface` implementation for the synchronous pass in `WhileOp` that allows arguments to be consumed.
* Modifications to `SingleThreadedExecutorImpl` to query the `CanConsumeArg()` method and consume arguments wherever possible.

Potential future additions:
* Add `ConsumeArg()` support to the default executor.
* Use custom call frames with `ConsumeArg()` support in more functional ops.

PiperOrigin-RevId: 309981157
Change-Id: I02f7b9f5611a27b087ee0058540ab7e38e70c21d