[TF MLIR SI] Don't constant fold, only consider result of folding
This results in less changes to the module during shape inference (e.g., only shapes are changed, no constant nodes are created). Effectively this computes the folded result and then just uses that information locally. Which is conceptually more wasteful (as a subsequent canonicalize pass may need to recompute these) but is less surprising and avoids dropping attributes during this part.

There is still additional changes that need to be made to avoid doing needless computations here, this mostly focuses on decreasing graph mutations.

PiperOrigin-RevId: 311539328
Change-Id: Ib6daa331c1e18a6d23463aa945c87e59d253708b