[XLA] Fix alternate memory allocation of conditional operands.
Consider the following flattened HLO schedule of a conditional:

1: a = fusion()
   true_computation:
2:    parameter = parameter(0)
3:    ...
4:    ...
   false_computation:
5:    parameter = parameter(0)
6:    ...
7:    ...
8: conditional = conditional(pred, a, a)
9: b = fusion(a)

When we had a tensor that was a conditional operand (e.g. "a" in the example),
we reserved the alternate memory for the entire 1-8 range. This meant that when
we tried to allocate inside the called computations of the conditional, the
offset we picked wasn't available since it would fall within the 1-8 range. This
CL now reserves the conditional until the parameter of the earliest called
computations (1-2 range).

To allow efficient use of alternate memory by avoiding a very large conditional
from claiming the offset for the entire called computation, the conditional
operand might die within the called computation, allowing other HLOs inside the
called computations to reclaim that alternate memory offset. This creates a
subtlety for subsequent uses of conditional operands (e.g. "a" is used by a
fusion at 9). These subsequent uses will force evictions (and then do another
prefetch). After optimization, the graph might look like the following:

  a (Alternate Mem) = fusion()
  cs0 = copy-start(a)  # Must evict a because the allocation may die within
                       # called computation.
  cd0 (Default Mem) = copy-done(cs0)
  true_computation:
    parameter (Alternate Mem) = parameter(0)
    ...
    # parameter's alternate memory allocation may die here and another tensor
    # might use the same offset.
  false_computation:
    parameter (Alternate Mem) = parameter(0)
    ...
    # parameter's alternate memory allocation may die here and another tensor
    # might use the same offset.
  conditional = conditional(pred, a, a)
  cs1 = copy-start(cd0)  # May prefetch the value back to alternate memory.
  cd1 (Alternate Mem) = copy-done(cs1)
  b = fusion(cd1)

PiperOrigin-RevId: 312182824
Change-Id: I3ff5d019025ef96ced1aed4f6d170df677273348