Limit FillOp custom folder to int and float types to avoid crash
DenseElementsAttr::getValue doesn't work for complex and string types and denseelementsAttr::get with Attribute only works for int and float types.

It is possible to handle complex types in the custom folder but not doing that now as complex types are less common and it would be easier to handle those once we have an attribute type for complex types.

PiperOrigin-RevId: 312687907
Change-Id: I4596e82d7b7e1d353bfb045b39b451785a7474e7