Fix complex & integer types in reduce operations
For tf.math.reduce_variance and tf.math.reduce_std we need to take the
absolute value before computing the square of the difference. This can
be skipped for non-complex values. For complex values we achieve this by
simply multiplying with the conjugate.

Further, we explicitly disallow integer types in these APIs since they
lead to imprecise calculations.

PiperOrigin-RevId: 313821668
Change-Id: I5eef00dcb1ecfcea0dd67af0cdc1f9785732a8ae