[TF] Add support for higher rank batch shape to keras.layers.Conv{,2D}.
The support is added to the base Conv layer, but the underlying 1D & 3D NN ops don't
yet support it.  But that's very easy to add now; so if this change
goes in, I can add support for nn_ops conv1d, conv3d, and keras Conv1D and
Conv3D layers all together.

PiperOrigin-RevId: 314331499
Change-Id: Id1f723993881c206d006711afd3fc7921459df11